Split _vcs module into a package with per-backend files
Reorganize _vcs.py (449 lines) into a package with one module
per VCS backend, improving navigability. Tests restructured to
mirror the src layout with shared helpers in conftest.py.
