+++ tests/_parser/_vcs/conftest.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from subprocess import TimeoutExpired  # noqa: S404
+
+
+def _raise_timeout(_process: object) -> None:
+    msg = "cmd"
+    raise TimeoutExpired(msg, 5)
+
+
+def _raise_file_not_found(_process: object) -> None:
+    raise FileNotFoundError
